Add files via upload
This Excel assignment was provided by the Anudip Foundation. It includes various tasks and exercises designed to test and improve proficiency in Excel. The exercises cover a range of topics, including data manipulation, use of formulas, data visualization, and other essential Excel functionalities. This assignment is a comprehensive tool for anyone looking to enhance their Excel skills and knowledge.